# api/openapi.yaml
openapi: 3.1.0
info:
  title: OneIssuer Identity and Client Management API
  version: 0.1.0-dev.4
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
  description: |-
    Implemented management/current-user JSON contract for a self-hosted, single-Issuer
    deployment. It covers health, current-user session and consent Grant management, and the
    administrator Users, Clients, Sessions, and Audit APIs.

    Browser authentication is hosted as HTML at `/login`, `/register`,
    `/logout`, and `/auth/complete`; those form interactions are intentionally
    not represented as JSON operations here. Discovery, JWKS, Authorize, Token,
    UserInfo, Refresh Token, Revocation, Introspection, and RP-Initiated Logout are
    implemented standard protocol surfaces but intentionally remain outside this
    management API description.

    Authenticated JSON calls use a server-side session cookie. State-changing
    calls additionally require the clear CSRF value returned in the
    `X-CSRF-Token` response header by `/api/v1/me`, `/api/v1/me/sessions`, `/api/v1/me/grants`,
    `/api/admin/v1/me`, or an administrative list endpoint. The browser also
    sends the matching CSRF cookie, and OneIssuer applies same-origin checks.
servers:
  - url: /
    description: Configured OneIssuer origin
...
paths:
...
  /api/v1/me:
    get:
      tags: [Current user]
      operationId: getCurrentUser
      summary: Read the current non-sensitive identity and session ID
      description: This is an account API, not the OIDC UserInfo endpoint.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current identity and session summary.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            X-CSRF-Token:
              $ref: "#/components/headers/CSRFToken"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /api/v1/me/sessions:
    get:
      tags: [Current user]
      operationId: listCurrentUserSessions
      summary: List only the current user's session summaries
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A bounded keyset-paginated session page.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            X-CSRF-Token:
              $ref: "#/components/headers/CSRFToken"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
  /api/v1/me/sessions/revoke-others:
    post:
      tags: [Current user]
      operationId: revokeOtherCurrentUserSessions
      summary: Revoke all of the current user's sessions except the current one
      security:
        - sessionCookie: []
          csrfToken: []
      responses:
        "200":
          description: Number of sessions newly revoked.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevokedCount"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /api/v1/me/grants:
    get:
      tags: [Current user]
      operationId: listCurrentUserConsentGrants
      summary: List the current user's client consent Grants
      description: |-
        Uses an owner-safe keyset cursor containing only `updated_at` and public
        `client_id`. Internal Grant, Client, family, and Session identifiers are absent.
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/GrantCursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A bounded owner-only Grant page.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            X-CSRF-Token:
              $ref: "#/components/headers/CSRFToken"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrantPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /api/v1/me/grants/revoke:
    post:
      tags: [Current user]
      operationId: revokeCurrentUserConsentGrant
      summary: Revoke one owned Grant by public client_id
      description: |-
        Atomically revokes the Grant, increments its authority version, and revokes
        every dependent Refresh family and live Access metadata row. Unknown Clients
        and another user's Grant are indistinguishable from not found. Repeating an
        already completed revoke is idempotent and does not append another Audit event.
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeGrantRequest"
      responses:
        "200":
          description: Safe current state of the revoked Grant.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedGrant"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /api/admin/v1/me:
    get:
      tags: [Administration]
      operationId: getCurrentAdministrator
      summary: Read the current administrator identity and session ID
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current active administrator and session summary.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            X-CSRF-Token:
              $ref: "#/components/headers/CSRFToken"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /api/admin/v1/users:
    get:
      tags: [Administration]
      operationId: listUsers
      summary: Search and list users with bounded keyset pagination
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/UserSearch"
      responses:
        "200":
          description: Credential-free user page.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            X-CSRF-Token:
              $ref: "#/components/headers/CSRFToken"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Administration]
      operationId: createUser
      summary: Create a managed user and initial password credential
      description: Creating an administrator also requires recent authentication.
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Credential-free created user.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/admin/v1/clients:
    get:
      tags: [Administration]
      operationId: listClients
      summary: List credential-free clients with keyset pagination
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Client page containing no clear secret or secret digest.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            X-CSRF-Token:
              $ref: "#/components/headers/CSRFToken"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Administration]
      operationId: createClient
      summary: Create a public or confidential client
      description: |-
        Requires recent authentication. A confidential clear secret appears
        only in this successful response; a public client has no secret.
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "201":
          description: Created client and, for a confidential client, its one-time secret.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedClientResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/admin/v1/sessions:
    get:
      tags: [Administration]
      operationId: listSessions
      summary: List safe login-session summaries
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Administrative session page.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            X-CSRF-Token:
              $ref: "#/components/headers/CSRFToken"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
  /api/admin/v1/audit-events:
    get:
      tags: [Administration]
      operationId: listAuditEvents
      summary: List immutable, value-free audit events
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/EventType"
      responses:
        "200":
          description: Append-only audit event page.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            X-CSRF-Token:
              $ref: "#/components/headers/CSRFToken"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: oneissuer_session
      description: |-
        Opaque server-side session cookie. The deployed name is configurable;
        production requires a Secure `__Host-` name. The database stores only a
        domain-separated digest. Never copy this value into logs or examples.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |-
        Clear double-submit value returned by an authenticated GET. Mutations
        also require the matching CSRF cookie and same-origin validation.
  headers:
    RequestID:
      description: Bounded caller-provided or server-generated request ID.
      schema:
        type: string
        pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
    CSRFToken:
      description: Rotated or reused clear CSRF value for a subsequent same-origin mutation. Treat as sensitive.
      schema:
        type: string
        writeOnly: true
        pattern: "^c1_[A-Za-z0-9_-]{43}$"
...
  schemas:
...
    MeResponse:
      type: object
      additionalProperties: false
      required: [user, session_id]
      properties:
        user:
          $ref: "#/components/schemas/User"
        session_id:
          type: string
          format: uuid
...
  responses:
...
    Unauthorized:
      description: Missing, invalid, expired, revoked, or disabled session (`authentication_required`).
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Safe internal failure with no database, stack, or credential detail (`internal_error`).
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...